:adhesive_bandage: starship: Fix separator colour after conflicting files
The separator in the "git status" module was inheriting foreground
colour from the previous displayed text, so the one right after the
conflicting files submodule was displayed red.

The fix simply forces the separator to use the default style of the "git
status" module, instead of inheriting from previous displays.