:recycle: hypr: Made workspace binds generic to allow combining number-based workspaces and other binds
Previously, the `traverse-workspaces` function (base helper for
iterating over workspaces) provided iteration over a sort of hardcoded
list: it was only mapping over the `lib.range 1 12` of workspaces,
providing no clear way to add other binds than these number-based
workspaces associations.

This refactor introduces a new list: `workspaces` (the old `workspaces`
binding which was just a range is renamed `workspaceRange`) which is a
list of pairs: (hyprlang KEY, hyprlang WORKSPACE).
The previous number-based binds are now fed into this list. The fact
that it is a separate list makes it easier to add non number-based binds
(such as a bind for `previous` for instance).
`traverse-workspaces` has been adapted to map over this list.